requirements.txt
reflex~=0.7.0
SQLAlchemy~=2.0.37
protobuf~=5.29.3
passlib~=1.7.4
pip~=24.2
wheel~=0.44.0
docutils~=0.21.2
Jinja2~=3.1.4
grpcio~=1.67.1
google-generativeai~=0.8.4
requests~=2.32.3
openai~=1.65.3
attrs~=25.1.0
distro~=1.9.0
keyring~=25.4.1